+++ views/feedback.py
@@ -1,4 +1,4 @@
-from flask import Blueprint, request, render_template, redirect, url_for, session
+from flask import Blueprint, request, render_template, redirect, url_for, session, flash
 import psycopg2 as db
 import os
 
@@ -40,7 +40,9 @@ def validate_feedback(form):
     # checking_service
     form_checking_service = form.get("checking_service", "").strip()
     if pointChecker(form_checking_service) == False:
-        form.errors["checking_service"] = "Checking Service support must be between 0 and 5."
+        form.errors[
+            "checking_service"
+        ] = "Checking Service support must be between 0 and 5."
     else:
         form.data["checking_service"] = form_checking_service
     # baggage_handling
@@ -71,22 +73,60 @@ def airline_feedback(ticker):
             "SELECT * FROM feedback WHERE airline_ticker = '{0}'".format(ticker)
         )
         feedbacks = cur.fetchall()
+        cur.execute("SELECT DISTINCT satisfaction FROM feedback ORDER BY satisfaction ")
+        satisfaction = cur.fetchall()
         cur.close()
         return render_template(
-            "feedback_airlines.html", feedbacks=feedbacks, ticker=ticker
+            "feedback_airlines.html", feedbacks=feedbacks, ticker=ticker, satisfaction = satisfaction
         )
-    else:
+    if 'satisfaction' in request.form and request.method == "POST":
+        cur.execute("SELECT DISTINCT satisfaction FROM feedback ORDER BY satisfaction ")
+        satisfaction = cur.fetchall()
+        satisfactionFilter = request.form["satisfaction"]
+        if satisfactionFilter == "*":
+            cur.execute(
+            "SELECT * FROM feedback WHERE airline_ticker = '{0}'".format(ticker)
+            )
+        else:
+            cur.execute(
+            (
+                "SELECT * FROM feedback WHERE satisfaction = %s and airline_ticker = %s"
+            ),
+            (satisfactionFilter, ticker),
+            )
+        feedbacks = cur.fetchall()
+        if(len(feedbacks)) == 0:
+            flash("No comment of exists!", "danger")
+        return render_template("feedback_airlines.html", feedbacks=feedbacks, ticker=ticker, satisfaction = satisfaction)
+
+    if request.method == "POST":
+        if not "id" in session:
+            flash("Please login!", "danger")
+            return redirect(url_for("user_authentication.login"))
         id_feedback = request.form.getlist("feedback_id")
+        i = 0
+        for form_feedback_key in id_feedback:
+            cur.execute("SELECT user_name FROM feedback WHERE id = {0}".format(form_feedback_key))
+            username = cur.fetchone()
+            if session["username"] != username[i]:
+                flash("You can only delete your own comment", "danger")
+                return redirect(url_for("feedback.airline_feedback", ticker=ticker))
+            i = i+1
+
+        cur.execute("SELECT DISTINCT satisfaction FROM feedback ORDER BY satisfaction ")
+        satisfaction = cur.fetchall()
+
         for form_feedback_key in id_feedback:
             cur.execute("DELETE FROM feedback WHERE id = {0}".format(form_feedback_key))
         connection.commit()
         cur.close()
-        return redirect(url_for("feedback.airline_feedback", ticker=ticker))
+        return redirect(url_for("feedback.airline_feedback", ticker=ticker, satisfaction = satisfaction))
 
 
 @feedback.route("/feed_back/add/<ticker>", methods=["GET", "POST"])
 def add_feedback(ticker):
     if not "id" in session:
+        flash("Please login!", "danger")
         return redirect(url_for("user_authentication.login"))
     connection = db.connect(os.getenv("DATABASE_URL"))
     cur = connection.cursor()
@@ -102,6 +142,10 @@ def add_feedback(ticker):
         }
         return render_template("feedback_add.html", ticker=ticker, values=values)
     if request.method == "POST":
+        if not "id" in session:
+            flash("Please login!", "danger")
+            return redirect(url_for("user_authentication.login"))
+
         valid = validate_feedback(request.form)
         if not valid:
             return render_template(
@@ -109,7 +153,7 @@ def add_feedback(ticker):
                 ticker=ticker,
                 values=request.form,
             )
-            
+
         type = request.form["type"]
         classt = request.form["class"]
         satisfaction = request.form["satisfaction"]
@@ -139,6 +183,7 @@ def add_feedback(ticker):
 @feedback.route("/feed_back/update/<ticker>/<id>", methods=["POST", "GET"])
 def update_feedback(id, ticker):
     if not "id" in session:
+        flash("Please login!", "danger")
         return redirect(url_for("user_authentication.login"))
     connection = db.connect(os.getenv("DATABASE_URL"))
     cur = connection.cursor()
@@ -147,6 +192,8 @@ def update_feedback(id, ticker):
         feedback_info = cur.fetchall()
         id = feedback_info[0][0]
         ticker = feedback_info[0][9]
+        name = ()
+        name = feedback_info[0][1]
         values = {
             "type": feedback_info[0][2],
             "class": feedback_info[0][3],
@@ -157,16 +204,23 @@ def update_feedback(id, ticker):
             "cleanliness": feedback_info[0][8],
         }
         cur.close()
-        return render_template("feedback_update.html", id = id , ticker=ticker, values = values)
+        return render_template(
+            "feedback_update.html", id=id, ticker=ticker, values=values, name = name
+        )
     if request.method == "POST":
+        cur.execute("SELECT user_name FROM feedback WHERE id = {0}".format(id))
+
+        name = cur.fetchone()
+        if session["username"] != name[0]:
+            flash("You can only edit your rating!", "danger")
+            return redirect(url_for("feedback.airline_feedback", ticker=ticker))
+
         valid = validate_feedback(request.form)
         if not valid:
             return render_template(
-                "feedback_update.html",
-                id = id,
-                ticker=ticker,
-                values=request.form
+                "feedback_update.html", id=id, ticker=ticker, values=request.form
             )
+
         type = request.form["type"]
         classt = request.form["class"]
         satisfaction = request.form["satisfaction"]
